+++ CopilotKit/packages/eslint-config-custom/CHANGELOG.md
@@ -1,5 +1,17 @@
 # eslint-config-custom
 
+## 0.6.0-mme-experimental-actions.1
+
+### Minor Changes
+
+- Alpha
+
+## 0.6.0-mme-experimental-actions.0
+
+### Minor Changes
+
+- QA new action type
+
 ## 0.5.1
 
 ### Patch Changes