+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,27 @@
 # ui
 
+## 0.21.0-mme-experimental-actions.1
+
+### Minor Changes
+
+- Alpha
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@0.5.0-mme-experimental-actions.1
+
+## 0.21.0-mme-experimental-actions.0
+
+### Minor Changes
+
+- QA new action type
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@0.5.0-mme-experimental-actions.0
+
 ## 0.20.1
 
 ### Patch Changes