+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,29 @@
 # ui
 
+## 0.31.0-mme-experimental-actions.1
+
+### Minor Changes
+
+- Alpha
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-core@0.21.0-mme-experimental-actions.1
+  - @copilotkit/shared@0.5.0-mme-experimental-actions.1
+
+## 0.31.0-mme-experimental-actions.0
+
+### Minor Changes
+
+- QA new action type
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-core@0.21.0-mme-experimental-actions.0
+  - @copilotkit/shared@0.5.0-mme-experimental-actions.0
+
 ## 0.30.1
 
 ### Patch Changes