+++ CopilotKit/packages/tailwind-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # tailwind-config
 
+## 0.5.0-mme-experimental-actions.1
+
+### Minor Changes
+
+- Alpha
+
+## 0.5.0-mme-experimental-actions.0
+
+### Minor Changes
+
+- QA new action type
+
 ## 0.4.1
 
 ### Patch Changes