+++ CopilotKit/packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,17 @@
 # tsconfig
 
+## 0.10.0-mme-experimental-actions.1
+
+### Minor Changes
+
+- Alpha
+
+## 0.10.0-mme-experimental-actions.0
+
+### Minor Changes
+
+- QA new action type
+
 ## 0.9.1
 
 ### Patch Changes